CustomLineItemDraft to interface. #942
