+++ sphere-models/src/it/java/io/sphere/sdk/carts/CustomLineItemFixtures.java
@@ -13,6 +13,6 @@ public class CustomLineItemFixtures {
 
     public static CustomLineItemDraft createCustomLineItemDraft(final TaxCategory taxCategory) {
         final MonetaryAmount money = MoneyImpl.of(new BigDecimal("23.50"), EUR);
-        return CustomLineItemDraft.of(en("thing"), "thing-slug", money, taxCategory, 5L);
+        return CustomLineItemDraft.of(en("thing"), "thing-slug", money, taxCategory, 5L, null);
     }
 }
